--- dipy/nn/cnn_1d_denoising/__init__.py
@@ -1,11 +1,0 @@
-from dipy.utils.optpkg import optional_package
-from packaging.version import parse
-
-tf, have_tf, _ = optional_package("tensorflow")
-layers, _, _ = optional_package("tensorflow.keras.layers")
-
-if have_tf:
-    if parse(tf.__version__) < parse('2.0.0'):
-        raise ImportError('Please upgrade to TensorFlow 2+')
-
-    from dipy.nn.cnn_1d_denoising.cnn_1d_denoising import Cnn1DDenoiser
