+++ dipy/nn/tests/test_cnn_1denoiser.py
@@ -2,12 +2,11 @@ import pytest
 import numpy as np
 from dipy.utils.optpkg import optional_package
 tf, have_tf, _ = optional_package('tensorflow')
-
 if have_tf:
     from dipy.nn.cnn_1d_denoising import Cnn1DDenoiser
 
 @pytest.mark.skipif(not have_tf, reason='Requires TensorFlow')
-def test_default_1CNN_sequential():
+def test_default_Cnn1DDenoiser_sequential():
     # Create dummy data    
     normal_img = np.random.rand(10, 10, 10, 30)
     nos_img = normal_img + np.random.normal(loc=0.0, scale=0.1, size=normal_img.shape)
@@ -23,7 +22,7 @@ def test_default_1CNN_sequential():
     assert data.shape == x.shape
 
 @pytest.mark.skipif(not have_tf, reason='Requires TensorFlow')
-def test_default_1CNN_flow(pytestconfig):
+def test_default_Cnn1DDenoiser_flow(pytestconfig):
     # Create dummy data
     normal_img = np.random.rand(10, 10, 10, 30)
     nos_img = normal_img + np.random.normal(loc=0.0, scale=0.1, size=normal_img.shape)
